Create Tweet.py
The device “TwitterBot” is built by using Raspberry Pi3 board and DHT sensor and is developed such that it tweets a refreshing morning quote at 4:00 AM every day and also measures the temperature of the room. If the temperature is less than 15 ºC then it turns off the AC and if more than 25 ºC it turns on the AC. Here AC is signified by a blue color LED. There are more than 1000 morning quotes defined in an external file(quotes.py) which is used by the program. The program chooses a quote each day
and posts it on the twitter account of the user at 4:00 AM. Time is monitored by the
program using python's time library. While performing the task of tweeting, the program
turns on the green LED which is connected to the GPIO24 pin of the Raspberry Pi3
board as an indicator that Bot is tweeting something. This program communicates with
the twitter API and makes a secure connection with the twitter server and accesses the
user account using the authentication keys which are defined in another external python
file(primary_key.py).